refactor(test): split the harness by responsibility
Keep the public quasar-test API unchanged while moving world state, setup, fixtures, and assertions into focused private modules and enabling its executable crate example.
