+++ test/Cargo.toml
@@ -24,4 +24,4 @@ quasar-test-derive = { path = "derive", version = "=0.1.0" }
 spl-token = { version = "=9.0.0", default-features = false, features = ["no-entrypoint"] }
 
 [lib]
-doctest = false
+doctest = true
